% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEnet.R
\name{calculate_standard_weights}
\alias{calculate_standard_weights}
\title{Calculate weights for standard adaptive elastic net}
\usage{
calculate_standard_weights(coefficients)
}
\arguments{
\item{coefficients}{Current coefficient estimates}
}
\value{
Vector of weights (1/|beta|)
}
\description{
Calculate weights for standard adaptive elastic net
}
\keyword{internal}
