% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEnet2.R
\name{cv_covariate_gcdnet}
\alias{cv_covariate_gcdnet}
\title{Cross-validation for covariate-dependent structure using gcdnet (Internal)}
\usage{
cv_covariate_gcdnet(
  X_std,
  y,
  standardize_X_gcdnet,
  lambda2_sequence,
  nlambda,
  lambda_selection_rule,
  gamma_sequence,
  prev_coefs_std,
  covariate_std,
  nvars,
  num_folds,
  num_cores,
  nobs,
  verbose
)
}
\description{
Cross-validation for covariate-dependent structure using gcdnet (Internal)
}
\keyword{internal}
