% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEnet2.R
\name{cv_lambda2}
\alias{cv_lambda2}
\title{Cross-validation across lambda2 values (Internal)}
\usage{
cv_lambda2(
  X_std,
  y,
  standardize_X_gcdnet,
  lambda2_sequence,
  nlambda,
  num_folds,
  num_cores,
  nobs,
  penalty_factor = NULL,
  lambda_selection_rule
)
}
\arguments{
\item{X_std}{Standardized design matrix}

\item{y}{Response vector}

\item{standardize_X_gcdnet}{Logical, passed to gcdnet's standardize argument.}

\item{lambda2_sequence}{Sequence of lambda2 values}

\item{nlambda}{Number of lambda values}

\item{num_folds}{Number of CV folds}

\item{num_cores}{Number of cores}

\item{nobs}{Number of observations}

\item{penalty_factor}{Penalty factors for gcdnet}

\item{lambda_selection_rule}{Rule for lambda selection}
}
\value{
List with best criterion value, best lambda, best lambda2.
}
\description{
Cross-validation across lambda2 values (Internal)
}
\keyword{internal}
