% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEnet.R
\name{fit_adaptive_net}
\alias{fit_adaptive_net}
\title{Fit Structure-Adaptive Elastic Net model (Internal)}
\usage{
fit_adaptive_net(
  y,
  X_std,
  structure_info,
  max_iterations,
  standardize_X_glmnet,
  alpha_sequence,
  nlambda,
  lambda_min_ratio,
  lambda_selection_rule,
  gamma_sequence,
  num_folds,
  num_cores,
  verbose,
  nvars,
  nobs
)
}
\arguments{
\item{y}{Response vector}

\item{X_std}{Standardized design matrix}

\item{structure_info}{External structural information}

\item{max_iterations}{Maximum number of iterations}

\item{standardize_X_glmnet}{Logical, passed to glmnet's standardize argument.
Note: X_std is already standardized if saenet's standardize=TRUE.
This controls glmnet's internal standardization. Usually FALSE if X_std is pre-standardized.}

\item{alpha_sequence}{Sequence of alpha values}

\item{nlambda}{Number of lambda values}

\item{lambda_min_ratio}{Ratio of min to max lambda}

\item{lambda_selection_rule}{Rule for lambda selection ("lambda.min" or "lambda.1se")}

\item{gamma_sequence}{Sequence of gamma values}

\item{num_folds}{Number of cross-validation folds}

\item{num_cores}{Number of cores for parallel computation}

\item{verbose}{Whether to print progress messages}

\item{nvars}{Number of predictor variables}
}
\value{
A list with model fitting results on the standardized scale.
}
\description{
Fit Structure-Adaptive Elastic Net model (Internal)
}
\keyword{internal}
