% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEnet2.R
\name{fit_adaptive_net_gcdnet}
\alias{fit_adaptive_net_gcdnet}
\title{Fit Structure-Adaptive Elastic Net model using gcdnet (Internal)}
\usage{
fit_adaptive_net_gcdnet(
  y,
  X_std,
  structure_info,
  max_iterations,
  standardize_X_gcdnet,
  lambda2_sequence,
  nlambda,
  lambda_selection_rule,
  gamma_sequence,
  num_folds,
  num_cores,
  verbose,
  nvars,
  nobs
)
}
\arguments{
\item{y}{Response vector}

\item{X_std}{Standardized design matrix}

\item{structure_info}{External structural information}

\item{max_iterations}{Maximum number of iterations}

\item{standardize_X_gcdnet}{Logical, passed to gcdnet's standardize argument.
Note: X_std is already standardized if saenet's standardize=TRUE.
This controls gcdnet's internal standardization. Usually FALSE if X_std is pre-standardized.}

\item{lambda2_sequence}{Sequence of lambda2 values}

\item{nlambda}{Number of lambda values}

\item{lambda_selection_rule}{Rule for lambda selection ("lambda.min", "lambda.1se", or "bic")}

\item{gamma_sequence}{Sequence of gamma values}

\item{num_folds}{Number of cross-validation folds}

\item{num_cores}{Number of cores for parallel computation}

\item{verbose}{Whether to print progress messages}

\item{nvars}{Number of predictor variables}

\item{nobs}{Number of observations}
}
\value{
A list with model fitting results on the standardized scale.
}
\description{
Fit Structure-Adaptive Elastic Net model using gcdnet (Internal)
}
\keyword{internal}
