% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEnet.R
\name{optimize_tau}
\alias{optimize_tau}
\title{Optimize tau parameters for covariate-dependent structure (Internal)}
\usage{
optimize_tau(
  gamma,
  current_coefficients,
  covariate_std,
  nvars,
  max_iter = 5000,
  verbose = FALSE
)
}
\description{
Optimize tau parameters for covariate-dependent structure (Internal)
}
\keyword{internal}
