% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.saenet}
\alias{print.saenet}
\title{Print a Structure-Adaptive Elastic Net Model}
\usage{
\method{print}{saenet}(x, ...)
}
\arguments{
\item{x}{A fitted SAEnet model object}

\item{...}{Further arguments (not used)}
}
\value{
Invisibly returns the input object
}
\description{
Prints a summary of a fitted Structure-Adaptive Elastic Net model.
}
