% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SAENET.R
\name{SAENET.nnet}
\alias{SAENET.nnet}
\title{Use a stacked autoencoder to pre-train a feed-forward neural network.}
\usage{
SAENET.nnet(h, X.train, target, nn.control = NULL)
}
\arguments{
\item{h}{The object returned from \code{SAENET.train()}}

\item{X.train}{A matrix of training data.}

\item{target}{A vector of target values. If given as a factor, classification
will be performed, otherwise if an integer or numeric vector is given
\code{neuralnet()} will default to regression}

\item{nn.control}{A named list with elements to be passed as control parameters to \code{neuralnet()}. Package defaults used if no values entered.}
}
\value{
An object of class \code{nn} which can be used with the \code{neuralnet} package as normal
}
\description{
Use a stacked autoencoder to pre-train a feed-forward neural network.
}

