% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEnet.R
\name{calculate_group_weights}
\alias{calculate_group_weights}
\title{Calculate weights for group-based adaptive elastic net}
\usage{
calculate_group_weights(coefficients, group_indices, num_groups)
}
\description{
Calculate weights for group-based adaptive elastic net
}
\keyword{internal}
