% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAEnet.R
\name{cv_alpha}
\alias{cv_alpha}
\title{Cross-validation across alpha values}
\usage{
cv_alpha(
  X_std,
  y,
  standardize_X,
  alpha_sequence,
  nlambda,
  lambda_min_ratio,
  num_folds,
  num_cores = 1,
  penalty_factor = NULL
)
}
\arguments{
\item{X_std}{Standardized design matrix}

\item{y}{Response vector}

\item{standardize_X}{Whether to standardize the predictors}

\item{alpha_sequence}{Sequence of alpha values to try}

\item{nlambda}{Number of lambda values}

\item{lambda_min_ratio}{Ratio of min to max lambda}

\item{num_folds}{Number of cross-validation folds}

\item{num_cores}{Number of cores for parallel computation}

\item{penalty_factor}{Optional vector of penalty factors}
}
\value{
A list with best CV error, lambda, and alpha
}
\description{
Cross-validation across alpha values
}
\keyword{internal}
